separate out target description
