+++ fpy2/backend/cpp/ops.py
@@ -1,5 +1,12 @@
 """
-cpp backend: operation type tables.
+cpp backend: per-op signature abstractions.
+
+This module defines the *shape* of the cpp backend's operator
+descriptions — the per-arity ``CppOp`` dataclasses and the
+``ScalarOpTable`` that groups them — without prescribing *which*
+operators are supported.  See :mod:`fpy2.backend.cpp.target` for the
+default target description (the set of operators and signatures the
+emitter dispatches against).
 
 Each primitive operation is parameterized by *argument C++ types*
 (:class:`CppScalar`) and an *active rounding context*
@@ -23,10 +30,6 @@ At an op site the emitter consults:
   slot.  On mismatch the emitter falls back to the
   all-active-context signature and inserts an explicit
   ``static_cast`` per operand.
-
-This mirrors :mod:`fpy2.backend.cpp.ops` in shape, but the output
-slot carries a full :class:`Context` (so RM-specific signatures are
-distinct) rather than a bare :class:`CppScalar`.
 """
 
 from __future__ import annotations
@@ -34,21 +37,8 @@ from __future__ import annotations
 import dataclasses
 from typing import TypeAlias
 
-from ...ast.fpyast import (
-    Abs, Acos, Acosh, Add, Asin, Asinh, Atan, Atan2, Atanh, Cbrt,
-    Ceil, Copysign, Cos, Cosh, Div, Erf, Erfc, Exp, Exp2, Expm1,
-    Expr, Fdim, Fma, Floor, Fmod, Hypot, Lgamma, Log, Log10, Log1p,
-    Log2, Logb, Mul, NearbyInt, Neg, Pow, Remainder, RoundInt,
-    Sin, Sinh, Sqrt, Sub, Tan, Tanh, Tgamma, Trunc,
-)
-from ...number import (
-    RM,
-    INTEGER,
-    SINT8, SINT16, SINT32, SINT64,
-    UINT8, UINT16, UINT32, UINT64,
-)
+from ...ast.fpyast import Expr
 from ...number.context.context import Context
-from ...number.context.ieee754 import IEEEContext
 
 from .storage import choose_storage_scalar
 from .types import CppScalar
@@ -149,196 +139,13 @@ TernaryOpTable: TypeAlias = dict[type[Expr], list[TernaryCppOp]]
 
 @dataclasses.dataclass
 class ScalarOpTable:
-    """Per-op-kind tables of supported C++ signatures."""
+    """Per-op-kind tables of supported C++ signatures.
+
+    A target description (see :mod:`fpy2.backend.cpp.target`)
+    populates one of these and hands it to the emitter; the emitter
+    looks up signatures by ``(op type, operand C++ types, active
+    rounding context)`` at each dispatch site.
+    """
     unary: UnaryOpTable
     binary: BinaryOpTable
     ternary: TernaryOpTable
-
-
-# ---------------------------------------------------------------------
-# Default tables.
-#
-# We enumerate same-context signatures (input C++ types taken from
-# ``choose_storage_scalar(ctx.format())``) for every native context
-# the cpp backend supports.  FP bases pair with each
-# ``fesetround``-supported rounding mode (RNE/RTZ/RTP/RTN) so context
-# equality at the dispatch site can match the active RM exactly.
-
-_FP_RMS = (RM.RNE, RM.RTZ, RM.RTP, RM.RTN)
-
-
-def _fp_ctxs() -> list[Context]:
-    """All FP32 / FP64 contexts the cpp backend dispatches against
-    — one per supported rounding mode."""
-    return [
-        IEEEContext(es, nbits, rm)
-        for (es, nbits) in ((8, 32), (11, 64))
-        for rm in _FP_RMS
-    ]
-
-
-def _int_ctxs() -> list[Context]:
-    """Integer contexts the cpp backend dispatches against.
-    ``INTEGER`` is the unbounded fallback — kept for backward compat
-    with programs that don't pin a width."""
-    return [
-        SINT8, SINT16, SINT32, SINT64,
-        UINT8, UINT16, UINT32, UINT64,
-        INTEGER,
-    ]
-
-
-def _all_arith_ctxs() -> list[Context]:
-    return _fp_ctxs() + _int_ctxs()
-
-
-def _ty_of(ctx: Context) -> CppScalar:
-    """Map a same-context signature's context to the C++ scalar that
-    its inputs (and output) carry."""
-    return choose_storage_scalar(ctx.format())
-
-
-def _fp_unary(name: str) -> list[UnaryCppOp]:
-    """Same-context FP-only unary signatures for one ``<cmath>``
-    function.  One sig per FP context (FP32 / FP64 × the four
-    supported rounding modes)."""
-    return [
-        UnaryCppOp(name, is_func=True, arg_ty=_ty_of(c), out_ctx=c)
-        for c in _fp_ctxs()
-    ]
-
-
-def _fp_binary(name: str) -> list[BinaryCppOp]:
-    """Same-context FP-only binary signatures for one ``<cmath>``
-    function (function-call form, not infix)."""
-    return [
-        BinaryCppOp(name, is_infix=False,
-                    in1_ty=_ty_of(c), in2_ty=_ty_of(c), out_ctx=c)
-        for c in _fp_ctxs()
-    ]
-
-
-def _fp_ternary(name: str) -> list[TernaryCppOp]:
-    """Same-context FP-only ternary signatures."""
-    return [
-        TernaryCppOp(name,
-                     in1_ty=_ty_of(c), in2_ty=_ty_of(c),
-                     in3_ty=_ty_of(c), out_ctx=c)
-        for c in _fp_ctxs()
-    ]
-
-
-# ``<cmath>`` unary functions defined for FP types only.  Each entry
-# pairs an FPy AST node with its C++ name; signatures cover both FP32
-# and FP64 across every supported rounding mode.
-_UNARY_CMATH = (
-    # Roots
-    (Sqrt, 'std::sqrt'),
-    (Cbrt, 'std::cbrt'),
-    # FP rounding to integer-valued FP
-    (Ceil, 'std::ceil'),
-    (Floor, 'std::floor'),
-    (NearbyInt, 'std::nearbyint'),
-    (RoundInt, 'std::round'),
-    (Trunc, 'std::trunc'),
-    # Trigonometric
-    (Sin, 'std::sin'), (Cos, 'std::cos'), (Tan, 'std::tan'),
-    (Asin, 'std::asin'), (Acos, 'std::acos'), (Atan, 'std::atan'),
-    # Hyperbolic
-    (Sinh, 'std::sinh'), (Cosh, 'std::cosh'), (Tanh, 'std::tanh'),
-    (Asinh, 'std::asinh'), (Acosh, 'std::acosh'), (Atanh, 'std::atanh'),
-    # Exp / log family
-    (Exp, 'std::exp'), (Exp2, 'std::exp2'), (Expm1, 'std::expm1'),
-    (Log, 'std::log'), (Log10, 'std::log10'),
-    (Log1p, 'std::log1p'), (Log2, 'std::log2'),
-    # Special functions
-    (Erf, 'std::erf'), (Erfc, 'std::erfc'),
-    (Lgamma, 'std::lgamma'), (Tgamma, 'std::tgamma'),
-    # Numerical data
-    (Logb, 'std::logb'),
-)
-
-
-_BINARY_CMATH = (
-    (Pow, 'std::pow'),
-    (Fmod, 'std::fmod'),
-    (Remainder, 'std::remainder'),
-    (Copysign, 'std::copysign'),
-    (Fdim, 'std::fdim'),
-    (Hypot, 'std::hypot'),
-    (Atan2, 'std::atan2'),
-)
-
-
-_TERNARY_CMATH = (
-    (Fma, 'std::fma'),
-)
-
-
-def _make_unary_table() -> UnaryOpTable:
-    fp = _fp_ctxs()
-    ints = _int_ctxs()
-    same = _all_arith_ctxs()
-    table: UnaryOpTable = {
-        Neg: [UnaryCppOp('-', is_func=False, arg_ty=_ty_of(c), out_ctx=c)
-              for c in same],
-        Abs: (
-            [UnaryCppOp('std::fabs', is_func=True,
-                        arg_ty=_ty_of(c), out_ctx=c)
-             for c in fp]
-            + [UnaryCppOp('std::abs', is_func=True,
-                          arg_ty=_ty_of(c), out_ctx=c)
-               for c in ints]
-        ),
-    }
-    for op_cls, name in _UNARY_CMATH:
-        table[op_cls] = _fp_unary(name)
-
-    # ``logb`` under an integer output context lowers to
-    # ``std::ilogb`` plus an explicit widening cast to the output's
-    # storage.  ``std::ilogb`` returns C ``int`` (≥ 16 bits, usually
-    # 32) — wider than any realistic binary exponent for the FP
-    # inputs the cpp backend supports, so casting to the output's
-    # storage type is value-preserving in practice.
-    table[Logb] = table[Logb] + [
-        UnaryCppOp(
-            'std::ilogb', is_func=True,
-            arg_ty=_ty_of(fp_ctx), out_ctx=int_ctx,
-            cast_out=True,
-        )
-        for fp_ctx in fp
-        for int_ctx in ints
-    ]
-    return table
-
-
-def _make_binary_table() -> BinaryOpTable:
-    same = _all_arith_ctxs()
-    table: BinaryOpTable = {
-        op_cls: [
-            BinaryCppOp(name, True, _ty_of(c), _ty_of(c), c)
-            for c in same
-        ]
-        for op_cls, name in (
-            (Add, '+'),
-            (Sub, '-'),
-            (Mul, '*'),
-            (Div, '/'),
-        )
-    }
-    for op_cls, name in _BINARY_CMATH:
-        table[op_cls] = _fp_binary(name)
-    return table
-
-
-def _make_ternary_table() -> TernaryOpTable:
-    return {op_cls: _fp_ternary(name) for op_cls, name in _TERNARY_CMATH}
-
-
-def make_op_table() -> ScalarOpTable:
-    """Build the default cpp op table."""
-    return ScalarOpTable(
-        unary=_make_unary_table(),
-        binary=_make_binary_table(),
-        ternary=_make_ternary_table(),
-    )
